+++ relay-server/src/services/processor/span/processing.rs
@@ -16,7 +16,6 @@ use relay_config::Config;
 use relay_dynamic_config::{
     CombinedMetricExtractionConfig, ErrorBoundary, GlobalConfig, ProjectConfig,
 };
-use relay_event_normalization::AiOperationTypeMap;
 use relay_event_normalization::span::ai::enrich_ai_span;
 use relay_event_normalization::{
     BorrowedSpanOpDefaults, ClientHints, CombinedMeasurementsConfig, FromUserAgentInfo,
@@ -239,8 +238,6 @@ struct NormalizeSpanConfig<'a> {
     measurements: Option<CombinedMeasurementsConfig<'a>>,
     /// Configuration for AI model cost calculation
     ai_model_costs: Option<&'a ModelCosts>,
-    /// Configuration to derive the `gen_ai.operation.type` field from other fields
-    ai_operation_type_map: Option<&'a AiOperationTypeMap>,
     /// The maximum length for names of custom measurements.
     ///
     /// Measurements with longer names are removed from the transaction event and replaced with a
@@ -285,7 +282,6 @@ impl<'a> NormalizeSpanConfig<'a> {
                 global_config.measurements.as_ref(),
             )),
             ai_model_costs: global_config.ai_model_costs.as_ref().ok(),
-            ai_operation_type_map: global_config.ai_operation_type_map.as_ref().ok(),
             max_name_and_unit_len: aggregator_config
                 .max_name_length
                 .saturating_sub(MeasurementsConfig::MEASUREMENT_MRI_OVERHEAD),
@@ -348,7 +344,6 @@ fn normalize(
         performance_score,
         measurements,
         ai_model_costs,
-        ai_operation_type_map,
         max_name_and_unit_len,
         tx_name_rules,
         user_agent,
@@ -456,7 +451,7 @@ fn normalize(
 
     normalize_performance_score(span, performance_score);
 
-    enrich_ai_span(span, ai_model_costs, ai_operation_type_map);
+    enrich_ai_span(span, ai_model_costs);
 
     tag_extraction::extract_measurements(span, is_mobile);
 
@@ -809,7 +804,6 @@ mod tests {
             performance_score: None,
             measurements: None,
             ai_model_costs: None,
-            ai_operation_type_map: None,
             max_name_and_unit_len: 200,
             tx_name_rules: &[],
             user_agent: None,